remove & rename: 불필요한 파일 삭제 및 파일 위치 조정
